+++ packages/expo-router/src/fork/useLinking.native.ts
@@ -16,6 +16,7 @@ import {
   type ParamListBase,
   useNavigationIndependentTree,
 } from '../react-navigation/native';
+import { ROOT_CHAIN } from '../react-navigation/routers/stateKeys';
 import { extractExpoPathFromURL } from './extractPathFromURL';
 
 type Options = LinkingOptions<ParamListBase>;
@@ -130,7 +131,7 @@ export function useLinking(
       const routeNode = store?.routeNode;
       return routeNode
         ? createSeededRootState(parsedState, routeNode)
-        : completeParsedState(parsedState);
+        : completeParsedState(parsedState, ROOT_CHAIN);
     };
 
     if (url != null) {